+++ TelegramBotAPI-extensions-api/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/api/send/SendDice.kt
@@ -6,19 +6,22 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.KeyboardMarkup
 import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.Chat
+import com.github.insanusmokrassar.TelegramBotAPI.types.dice.DiceAnimationType
 
 suspend fun RequestsExecutor.sendDice(
     chatId: ChatIdentifier,
+    animationType: DiceAnimationType? = null,
     disableNotification: Boolean = false,
     replyToMessageId: MessageIdentifier? = null,
     replyMarkup: KeyboardMarkup? = null
 ) = execute(
-    SendDice(chatId, disableNotification, replyToMessageId, replyMarkup)
+    SendDice(chatId, animationType, disableNotification, replyToMessageId, replyMarkup)
 )
 
 suspend fun RequestsExecutor.sendDice(
     chat: Chat,
+    animationType: DiceAnimationType? = null,
     disableNotification: Boolean = false,
     replyToMessageId: MessageIdentifier? = null,
     replyMarkup: KeyboardMarkup? = null
-) = sendDice(chat.id, disableNotification, replyToMessageId, replyMarkup)
+) = sendDice(chat.id, animationType, disableNotification, replyToMessageId, replyMarkup)