+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/textsources/TextMentionTextSource.kt
@@ -2,16 +2,16 @@ package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsourc
 
 import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.MultilevelTextSource
 import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextPart
-import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.PrivateChat
+import com.github.insanusmokrassar.TelegramBotAPI.types.User
 import com.github.insanusmokrassar.TelegramBotAPI.utils.*
 
 class TextMentionTextSource(
     override val source: String,
-    privateChat: PrivateChat,
+    val user: User,
     textParts: List<TextPart>
 ) : MultilevelTextSource {
     override val textParts: List<TextPart> by lazy { source.fullListOfSubSource(textParts) }
-    override val asMarkdownSource: String by lazy { source.textMentionMarkdown(privateChat.id) }
-    override val asMarkdownV2Source: String by lazy { textMentionMarkdownV2(privateChat.id) }
-    override val asHtmlSource: String by lazy { textMentionHTML(privateChat.id) }
+    override val asMarkdownSource: String by lazy { source.textMentionMarkdown(user.id) }
+    override val asMarkdownV2Source: String by lazy { textMentionMarkdownV2(user.id) }
+    override val asHtmlSource: String by lazy { textMentionHTML(user.id) }
 }